pyproject.toml
[project]
name = "arch-autopilot"
version = "0.1.0"
description = ""
authors = [
    {name = "ChetanPatil201",email = "chetanp201jb@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langgraph (>=1.0.5,<2.0.0)",
    "langchain-openai (>=1.1.6,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-hcl2 (>=7.3.1,<8.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[project.scripts]
arch-autopilot = "arch_autopilot.cli:main"

[tool.poetry]
packages = [{include = "arch_autopilot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
